backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
flask-cors==4.0.0
asyncpg==0.29.0
SQLAlchemy[asyncio]==2.0.23
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.1.0.post1
python-jose==3.3.0